chore: refactor folder structure (#131)
