+++ src/pages/[organization]/users/index.tsx
@@ -55,7 +55,7 @@ import { useTranslation } from 'react-i18next'
 import { AbilityContext } from 'src/layouts/components/acl/Can'
 
 // ** Hooks Imports
-import { useCurrentOrganization } from 'src/hooks/useCurrentOrganization'
+import { useCurrentOrganization } from 'src/hooks'
 
 interface CellType {
   row: OrganizationUserResponseDto
