+++ src/pages/organization/new/index.tsx
@@ -76,7 +76,7 @@ const CreateOrganizationPage = () => {
 
   return (
     <Box className='content-center'>
-      <Card sx={{ width: 1 / 2 }}>
+      <Card sx={{ width: { lg: '40%', md: '50%', sm: '70%', xs: '100%' } }}>
         <CardHeader title={t('new.create_organization')} />
         <CardContent>
           <form onSubmit={handleSubmit(onSubmit)}>
