+++ src/utils/router/organization.ts
@@ -3,7 +3,7 @@ export const getSelectOrganizationUrl = (): string => {
 }
 
 export const getCreateOrganizationUrl = (): string => {
-  return '/organization/new'
+  return '/organization/new/'
 }
 
 export const getOrganizationDefaultHomeUrl = (orgUniqueName: string | undefined): string => {
